% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplication.R
\name{get_deduplicated_version}
\alias{get_deduplicated_version}
\title{Extract deduplicated version from duplication objects}
\usage{
get_deduplicated_version(duplication)
}
\arguments{
\item{duplication}{the duplication object to be processed}
}
\value{
a text vector
}
\description{
This method extracts the deduplicated text vector of the input duplication object.
}
