% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplication.R
\name{print.duplication}
\alias{print.duplication}
\title{Display duplication objects}
\usage{
\method{print}{duplication}(duplication)
}
\arguments{
\item{duplication}{the duplication object to be displayed}
}
\value{
nothing
}
\description{
This method displays a preview of the input duplication object.
}
